store/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.2.2
Flask-Migrate==4.1.0
Flask-Session==0.5.0
WTForms>=3.0
python-dotenv>=1.0
paypal-checkout-serversdk>=1.0.1
redis>=5.0.0
cloudinary>=1.39.0
Pillow>=10.0.0
email-validator>=2.1.0
gunicorn>=21.2.0
